fix: refactor actions and test
