+++ tests/Feature/OrderProductTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature;
 
-use App\Actions\Order\StoreOrderActions;
+use App\Actions\Order\StoreOrderAction;
 use App\Actions\OrderProduct\StoreUpdateOrderProductActions;
 use App\Models\Product;
 use App\Models\User;
@@ -17,7 +17,7 @@ class OrderProductTest extends TestCase
     public function test_create_action_order_product()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
         $product = Product::factory()->create();
         $orderProduct = StoreUpdateOrderProductActions::execute($order, $product);
