+++ tests/Feature/OrderTest.php
@@ -2,10 +2,9 @@
 
 namespace Tests\Feature;
 
-use App\Actions\Order\PayOrderActions;
-use App\Actions\Order\StoreOrderActions;
-use App\Actions\Order\UpdateOrderActions;
-use App\Actions\Order\UpdatePayOrderActions;
+use App\Actions\Order\PayOrderAction;
+use App\Actions\Order\StoreOrderAction;
+use App\Actions\Order\UpdateOrderAction;
 use App\Actions\OrderProduct\StoreUpdateOrderProductActions;
 use App\Constants\Constants;
 use App\Models\Product;
@@ -29,7 +28,7 @@ class OrderTest extends TestCase
     public function test_edit_order_screen_can_be_rendered()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
         $user = User::factory()->create();
 
         $response = $this->actingAs($user)->get(route('orders.edit', $order));
@@ -39,7 +38,7 @@ class OrderTest extends TestCase
     public function test_order_pay_screen_can_be_rendered()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
         $response = $this->actingAs($user)->post(route('orders.orderPay', $order));
 
         $this->assertEquals("http://localhost", $response->getTargetUrl());
@@ -49,7 +48,7 @@ class OrderTest extends TestCase
     public function test_show_order_screen_can_be_rendered()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
 
         $arrayPay = [
             'reference' => $order->id,
@@ -66,7 +65,7 @@ class OrderTest extends TestCase
         ];
 
         $paymentGeteway = new Placetopay();
-        PayOrderActions::execute($arrayPay, $dataOrder, $paymentGeteway);
+        PayOrderAction::execute($arrayPay, $dataOrder, $paymentGeteway);
 
         $response = $this->actingAs($user)->get('orders/showOrder/'.$order->id);
         $response->assertStatus(200);
@@ -76,7 +75,7 @@ class OrderTest extends TestCase
     public function test_create_action_order()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
 
         $this->assertSame((int)$order->total, 1);
         $this->assertSame((string)$order->status, "CREATED");
@@ -86,7 +85,7 @@ class OrderTest extends TestCase
     public function test_update_action_order()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
         $product = Product::factory()->create();
         $orderProduct = StoreUpdateOrderProductActions::execute($order, $product);
 
@@ -99,14 +98,14 @@ class OrderTest extends TestCase
             'status' => $statusCreate,
         ];
 
-        $order = UpdateOrderActions::execute($order, $data);
+        $order = UpdateOrderAction::execute($order, $data);
         $this->assertTrue($order);
     }
 
     public function test_pay_order_action_order()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
 
         $arrayPay = [
             'reference' => $order->id,
@@ -123,7 +122,7 @@ class OrderTest extends TestCase
         ];
 
         $paymentGeteway = new Placetopay();
-        $response = PayOrderActions::execute($arrayPay, $dataOrder, $paymentGeteway);
+        $response = PayOrderAction::execute($arrayPay, $dataOrder, $paymentGeteway);
 
         $urlRedirect = substr($response->getTargetUrl(),0,43);
         $this->assertEquals( Constants::URL_REDIRECT_PLACETOPAY, $urlRedirect);
