+++ tests/Feature/ProductTest.php
@@ -2,11 +2,9 @@
 
 namespace Tests\Feature;
 
-use App\Actions\Order\StoreOrderActions;
+
 use App\Models\Product;
 use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class ProductTest extends TestCase
