+++ tests/Feature/RequestTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature;
 
-use App\Actions\Order\StoreOrderActions;
+use App\Actions\Order\StoreOrderAction;
 use App\Actions\Request\StoreRequestActions;
 use App\Constants\Constants;
 use App\Models\User;
@@ -17,7 +17,7 @@ class RequestTest extends TestCase
     public function test_store_request_actions()
     {
         $user = User::factory()->create();
-        $order = StoreOrderActions::execute($user->id);
+        $order = StoreOrderAction::execute($user->id);
         $arrayPay = [
             'reference' => 1,
             'total' => 1000,
